replace float with double

+++ logistic_regression.c
@@ -10,8 +10,8 @@
 
 typedef struct 
 {
-    float *w;
-    float b;
+    double *w;
+    double b;
     size_t in_dim;
     size_t out_dim;    
 } model;
@@ -20,8 +20,8 @@ typedef model gradient;
 
 struct metrics
 {
-    float loss;
-    float accuracy;
+    double loss;
+    double accuracy;
 };
 
 typedef struct
@@ -55,7 +55,7 @@ void initialize_weights_(const unsigned int seed, model *m)
 {
     srand(seed);
     
-    m->w = malloc(m->in_dim * sizeof(float));
+    m->w = malloc(m->in_dim * sizeof(double));
 
     for (size_t i = 0; i < m->in_dim; i++)
     {
@@ -65,38 +65,38 @@ void initialize_weights_(const unsigned int seed, model *m)
 }
 
 
-float sigmoid(float x)
+double sigmoid(double x)
 {
     return 1.0f / (1.0f + expf(-x));
 }
 
 
-float binary_crossentropy_loss(float y, float y_pred)
+double binary_crossentropy_loss(double y, double y_pred)
 {
     return -((y * logf(y_pred)) + ((1.0f - y) * logf(1.0f - y_pred)));
 }
 
 
 // Runs backpropagation for a single instance
-grad_and_metrics *backprop(model *m, float *x, float y)
+grad_and_metrics *backprop(model *m, double *x, double y)
 {
     struct metrics *metrics = (struct metrics *) malloc(sizeof(struct metrics *));
     gradient *grad = (gradient *) malloc(sizeof(gradient));
     grad->in_dim = m->in_dim;
     grad->out_dim = m->out_dim;
     
-    float z, a;     // z: logits    a: activations
+    double z, a;     // z: logits    a: activations
     z = vector_dot_product(m->w, x, m->in_dim);
     z += m->b;
     a = sigmoid(z);
 
     // Compute loss
-    float loss = binary_crossentropy_loss(y, a);
+    double loss = binary_crossentropy_loss(y, a);
 
     // Compute gradients
-    float dloss_dz = a - y;
-    float *grad_loss_wrt_w = vector_scalar_product(x, dloss_dz, m->in_dim);
-    float grad_loss_wrt_b = dloss_dz;
+    double dloss_dz = a - y;
+    double *grad_loss_wrt_w = vector_scalar_product(x, dloss_dz, m->in_dim);
+    double grad_loss_wrt_b = dloss_dz;
 
     grad->w = grad_loss_wrt_w;
     grad->b = grad_loss_wrt_b;
@@ -127,13 +127,13 @@ static inline void divide_grad_by_batch_size_(gradient *accum_grad,
 {
     for (size_t i = 0; i < accum_grad->in_dim; i++)
     {
-        accum_grad->w[i] /= (float) batch_size;
+        accum_grad->w[i] /= (double) batch_size;
     }
-    accum_grad->b /= (float) batch_size;
+    accum_grad->b /= (double) batch_size;
 }
 
 
-static inline void update_weights_(model *m, gradient *grad, float learning_rate)
+static inline void update_weights_(model *m, gradient *grad, double learning_rate)
 {
     for (size_t i = 0; i < m->in_dim; i++)
     {
@@ -143,7 +143,7 @@ static inline void update_weights_(model *m, gradient *grad, float learning_rate
 }
 
 
-struct metrics *train_step(model *m, float *x, float *y, float learning_rate,
+struct metrics *train_step(model *m, double *x, double *y, double learning_rate,
         const size_t batch_size)
 {
     struct metrics *metrics = malloc(sizeof(struct metrics));
@@ -154,7 +154,7 @@ struct metrics *train_step(model *m, float *x, float *y, float learning_rate,
     model *accum_grad = malloc(sizeof(model));
     accum_grad->in_dim = m->in_dim;
     accum_grad->out_dim = m->out_dim;
-    accum_grad->w = calloc(accum_grad->in_dim, sizeof(float));
+    accum_grad->w = calloc(accum_grad->in_dim, sizeof(double));
     accum_grad->b = 0.0f;
 
     grad_and_metrics *result;
@@ -182,7 +182,7 @@ struct metrics *train_step(model *m, float *x, float *y, float learning_rate,
 }
 
 
-void fit(model *m, dataset *ds, float learning_rate, const size_t batch_size,
+void fit(model *m, dataset *ds, double learning_rate, const size_t batch_size,
         const size_t n_epochs)
 {
     size_t n_batches = ds->n_samples / batch_size;
